Extract validateRedirectUrl into RedirectUrlValidationTrait to avoid duplication
